Use std::mutex in thread pool
Use std::mutex instead of lcos::no_mutex. no_mutex is "more correct" in the
sense that no lock actually needs to be taken at this point, but it's an
unnecessary optimization as the destructor will be called rarely. By using
std::mutex we avoid depending on higher level functionality (lcos) in the
thread pools.