"Fix" suspend_thread_timed test after changes to timed_executor
Changes to timed_executor to eagerly create timed threads lead to the
suspend_thread_timed segfaulting for schedulers that don't support suspension.
This happened because when an exception is thrown trying to suspend a worker
thread, the thread pool gets shut down immediately, but the timed threads are
still waiting to spawn. When they wake up and try to schedule a thread on the
thread pool the thread pool is gone and the test segfaults.