Make parallel_executor use thread_pool_executor helpers for bulk execution
Use regular, (potentially) hierarchical spawning, with tasks assigned to
cores in a chunked order, i.e. the first k tasks use a hint for the
first core, the next k tasks use a hint for the second core, and the
last k tasks use a hint for the last core. This is the same behaviour as
thread_pool_executor.