Add simple CPU thread pool-specific P1897 when_all and fork algorithms
fork is not proposed in P1897. It's used here as an alternative to a
shared_future, i.e. it allows splitting a DAG. It starts all work
eagerly until the predecessor sender, and returns a new sender that
spawns new tasks from that point onwards (not from the root sender that
was passed to fork). This may be possible to implement with
ensure_started instead.